migrated to PG

+++ db/migrate/20131206174958_add_contact_to_users.rb
@@ -1,5 +1,5 @@
 class AddContactToUsers < ActiveRecord::Migration
   def change
-    add_column :users, :contact, :integer
+    add_column :users, :contact, :bigint
   end
 end
